changes project structure, extracted NotationInstruction component
