+++ src/components/Board/props.ts
@@ -1,0 +1,5 @@
+export interface BoardProps {
+  playAsBlack?: boolean;
+  lightSquareColor?: string;
+  darkSquareColor?: string;
+}
