+++ src/components/NotationInstruction/props.ts
@@ -1,0 +1,3 @@
+export type NotationInstructionProps = {
+
+};
